Fix freeRegister set bug
And add a check method to recognize such bugs faster

+++ src/main/java/minijava/ir/assembler/instructions/ConstArgument.java
@@ -7,7 +7,7 @@ public class ConstArgument extends Argument {
 
   public final String value;
 
-  public ConstArgument(Register.Width width, int value) {
+  public ConstArgument(Register.Width width, long value) {
     super(width);
     this.value = "$" + value;
   }
